refactor: App.jsxからuseHabitActionsとuseBackupManagerを分離 (#221)
- useHabitActions: toggleHabit/handleUndo/addHabit/updateHabit/deleteHabit/archiveHabit/restoreHabit/handleDragEndを抽出
- useBackupManager: handleExport/handleImportClick/handleFileChange/handleImportConfirmを抽出
- useModalStateは既存実装を活用（変更なし）
- App.jsxを178行削減（221行→43行純増）
- 既存の動作・UIは一切変更なし
